code/product-service-graphql/requirements.txt
strawberry-graphql
typer
libcst==1.2.0
strawberry-graphql[debug-server]